requirements.txt
asgiref==3.3.1
beautifulsoup4==4.10.0
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-distill==2.4
django-meta==2.0.0
django-tinymce==3.2.0
idna==2.10
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.3
